% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseorder.R
\name{ecs_rdpurchaseorder}
\alias{ecs_rdpurchaseorder}
\title{销售出库处理}
\usage{
ecs_rdpurchaseorder(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
销售出库处理
}
\examples{
ecs_rdpurchaseorder()
}
